Move QueryContext and createQueryContext to their own file, queryContext.ts
