+++ ts/packages/agents/spelunker/src/searchCode.ts
@@ -5,17 +5,14 @@ import * as fs from "fs";
 import * as path from "path";
 
 import * as sqlite from "better-sqlite3";
-import { createJsonTranslator, Result, TypeChatJsonTranslator } from "typechat";
-import { createTypeScriptJsonValidator } from "typechat/ts";
+import { Result, TypeChatJsonTranslator } from "typechat";
 
-import { ChatModel, openai, TextEmbeddingModel } from "aiclient";
 import { createLimiter } from "common-utils";
 import { ActionResult, Entity } from "@typeagent/agent-sdk";
 import {
     createActionResultFromMarkdownDisplay,
     createActionResultFromError,
 } from "@typeagent/agent-sdk/helpers/action";
-import { loadSchema } from "typeagent";
 
 import {
     Blob,
@@ -25,99 +22,14 @@ import {
     ChunkId,
 } from "./chunkSchema.js";
 import { createDatabase, purgeFile } from "./databaseUtils.js";
-import {
-    makeEmbeddingModel,
-    loadEmbeddings,
-    preSelectChunks,
-} from "./embeddings.js";
+import { loadEmbeddings, preSelectChunks } from "./embeddings.js";
 import { console_log, resetEpoch } from "./logging.js";
-import { OracleSpecs } from "./oracleSchema.js";
 import { chunkifyPythonFiles } from "./pythonChunker.js";
+import { createQueryContext } from "./queryContext.js";
 import { retryOn429 } from "./retryLogic.js";
 import { ChunkDescription, SelectorSpecs } from "./selectorSchema.js";
 import { SpelunkerContext } from "./spelunkerActionHandler.js";
-import { SummarizerSpecs } from "./summarizerSchema.js";
 import { chunkifyTypeScriptFiles } from "./typescriptChunker.js";
-
-export interface QueryContext {
-    chatModel: ChatModel;
-    miniModel: ChatModel;
-    embeddingModel: TextEmbeddingModel;
-    oracle: TypeChatJsonTranslator<OracleSpecs>;
-    chunkSelector: TypeChatJsonTranslator<SelectorSpecs>;
-    chunkSummarizer: TypeChatJsonTranslator<SummarizerSpecs>;
-    databaseLocation: string;
-    database: sqlite.Database | undefined;
-}
-
-function captureTokenStats(req: any, response: any): void {
-    const inputTokens = response.usage.prompt_tokens;
-    const outputTokens = response.usage.completion_tokens;
-    const cost = inputTokens * 0.000005 + outputTokens * 0.000015;
-    console_log(
-        `    [Tokens used: prompt=${inputTokens}, ` +
-            `completion=${outputTokens}, ` +
-            `cost=\$${cost.toFixed(2)}]`,
-    );
-}
-
-function createQueryContext(): QueryContext {
-    const chatModel = openai.createChatModelDefault("spelunkerChat");
-    chatModel.completionCallback = captureTokenStats;
-    chatModel.retryMaxAttempts = 0;
-
-    const miniModel = openai.createChatModel(
-        undefined, // "GPT_4_O_MINI" is slower than default model?!
-        undefined,
-        undefined,
-        ["spelunkerMini"],
-    );
-    miniModel.completionCallback = captureTokenStats;
-    miniModel.retryMaxAttempts = 0;
-
-    const embeddingModel = makeEmbeddingModel();
-
-    const oracle = createTranslator<OracleSpecs>(
-        chatModel,
-        "oracleSchema.ts",
-        "OracleSpecs",
-    );
-    const chunkSelector = createTranslator<SelectorSpecs>(
-        miniModel,
-        "selectorSchema.ts",
-        "SelectorSpecs",
-    );
-    const chunkSummarizer = createTranslator<SummarizerSpecs>(
-        miniModel,
-        "summarizerSchema.ts",
-        "SummarizerSpecs",
-    );
-
-    const databaseFolder = path.join(
-        process.env.HOME ?? "",
-        ".typeagent",
-        "agents",
-        "spelunker",
-    );
-    const mkdirOptions: fs.MakeDirectoryOptions = {
-        recursive: true,
-        mode: 0o700,
-    };
-    fs.mkdirSync(databaseFolder, mkdirOptions);
-
-    const databaseLocation = path.join(databaseFolder, "codeSearchDatabase.db");
-    const database = undefined;
-    return {
-        chatModel,
-        miniModel,
-        embeddingModel,
-        oracle,
-        chunkSelector,
-        chunkSummarizer,
-        databaseLocation,
-        database,
-    };
-}
 
 // Answer a question; called from request and from searchCode action
 export async function searchCode(
@@ -430,17 +342,6 @@ export function prepareSummaries(db: sqlite.Database): string {
         lines.push(summaryRow.signature);
     }
     return lines.join("\n");
-}
-
-function createTranslator<T extends object>(
-    model: ChatModel,
-    schemaFile: string,
-    typeName: string,
-): TypeChatJsonTranslator<T> {
-    const schema = loadSchema([schemaFile], import.meta.url);
-    const validator = createTypeScriptJsonValidator<T>(schema, typeName);
-    const translator = createJsonTranslator<T>(model, validator);
-    return translator;
 }
 
 export interface FileMtimeSize {
